support extension module reduce
